changing state changing operations in event sink ui to post

+++ components/mediation-ui/org.wso2.carbon.event.sink.config.ui/src/main/resources/web/event-sink-config/update_event_sink_ajaxprocessor.jsp
@@ -67,6 +67,11 @@
             }
         }
     } else if (action.equals("delete")) {
+        //ignore methods other than post
+        if (!request.getMethod().equalsIgnoreCase("POST")) {
+            response.sendError(405);
+            return;
+        }
         String name = request.getParameter("name");
         if (!(publishEventMediatorConfigAdminClient.deleteEventSink(name))) {
             responseText="false";
